+++ cobaya/bib.py
@@ -74,6 +74,7 @@ def bib_script():
     # Parse arguments and launch
     import argparse
     parser = argparse.ArgumentParser(
+        prog="cobaya bib",
         description="Prints bibliography to be cited for a component or input file.")
     parser.add_argument("components_or_files", action="store", nargs="+",
                         metavar="component_name or input_file.yaml",