+++ CHANGELOG.md
@@ -14,6 +14,11 @@
 - `[prefix].updated.yaml` is now `[prefix].minimize.updated.yaml` (GetDist needs to know the original sampler).
 - Loads covmat correcly when starting from PolyChord sample.
 
+### Collections
+
+- Collections are picklable again.
+- Slices with ommited limits, e.g. `[::2]`, now work.
+
 
 ## 3.0  – 2020-05-12